Move context map builders to separate AST namespace
