+++ src/Dto/Parsing/ContextMapBuilderInterface.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the TODO Registrar project.
+ *
+ * (c) Anatoliy Melnikov <5785276@gmail.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Aeliot\TodoRegistrar\Dto\Parsing;
+
+use Aeliot\TodoRegistrarContracts\ContextNodeInterface;
+
+interface ContextMapBuilderInterface
+{
+    /**
+     * @return array<int, list<ContextNodeInterface>>
+     */
+    public function buildContextMap(): array;
+}
